Refactor input into new file.
